examples/react-crewai-market-research/requirements.txt
# CrewAI and dependencies
crewai>=0.152.0
crewai-tools>=0.8.0

# LLM providers
openai>=1.0.0

# HTTP and API clients
requests>=2.31.0
httpx>=0.25.0

# Environment management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Optional: For local LLM support
# ollama>=0.1.0

# Optional: For additional tools
# beautifulsoup4>=4.12.0  # Web scraping
# PyPDF2>=3.0.0           # PDF reading
# pandas>=2.0.0           # Data analysis
